Implemented writing of all detector log fields to database in DatabaseLogSink now including ublox_counter_progress and deadtime_factor. The deadtime_factor should be a good indicator for the registered events compared to all events seen by the ublox. Changed ublox_counter_progress to long int.

+++ source/cluster/include/detector.h
@@ -119,7 +119,7 @@ private:
 	
 	DetectorLog::Data m_current_data;
 	std::uint16_t m_last_ublox_counter {};
-	int m_current_ublox_progress {};
+	long int m_current_ublox_progress {};
 	
 	Ringbuffer<double, 100> m_pulselength {};
